// src/app/models/parent.ts
export class Parent {
  pid: number;
  org: number;
  title1: string;
  lname1: string;
  fname1: string;
  email1: string;
  title2: string;
  lname2: string;
  fname2: string;
  email2: string;
  address1: string;
  address2: string;
  city: string;
  state: string;
  zip: string;
  comment: string;
}